Fixes #21
This is an important commit because it changes the way scone works
with run=TRUE and eval=TRUE. Previously, eval=TRUE meant that the
normalized data were not returned and only the evaluations were.

This doesn't make much sense, since we have to normalize the data
to be able to evaluate them.